+++ lib/spawn.cc
@@ -250,7 +250,7 @@ spawn(spawn_descriptor *sd)
 	e.te_arg[2] = start_env->user_grant;
     } else {
 	e.te_arg[0] = 0;
-	e.te_arg[1] = (uint64_t) spawn_env_va;
+	e.te_arg[1] = (uintptr_t) spawn_env_va;
     }
     error_check(sys_thread_start(tobj, &e,
 				 thread_label.to_ulabel(),
